MDocValidator filename and format change
